+++ src/views/Dashboard/Explore/interfaces/ICoreGarden.json
