+++ src/views/Dashboard/Explore/interfaces/IStrategyGarden.json
